se agrego nueva variable en destroy de la controladora Categorias

+++ app/Http/Controllers/Admin/CategoriaController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
 use App\Models\Categoria;
+use App\Models\Movimiento;
 use Illuminate\Http\Request;
 use Illuminate\Database\QueryException; 
 
@@ -102,11 +103,21 @@ class CategoriaController extends Controller
     public function destroy(Request $request, $id)
     {
         try {
-           
-            Categoria::destroy($id);
+           //se creo una variable nueva solo para eliminar categorias que no esten en uso
+            $movimientos= Movimiento::where('id_categoria', $id)->count();
+            if ($movimientos == 0) {
+                Categoria::destroy($id);
+                session()->flash('alert-success', trans('message.successaction'));
+            }
 
-            session()->flash('alert-success', trans('message.successaction'));
+            else{
+                   session()->flash('alert-danger', '¡Error! La categoria está siendo utilizada, no se puede eliminar');
+            }
+
+           
             return redirect()->route('admin.categorias.index');
+
+            
         } catch (QueryException  $ex) {
             session()->flash('alert-danger', $ex->getMessage());
             return redirect()->route('admin.categorias.index');
